feat: Update App component to redirect user based on authentication status
This commit updates the App component to include a useEffect hook that redirects the user to the appropriate page based on their authentication status. If the user is authenticated, they will be redirected to the "/list_users" page. Otherwise, they will be redirected to the "/login" page. This ensures that the user is directed to the correct page upon loading the application.